// frontend/src/types/auth.types.ts
export type Role = 'CUSTOMER' | 'SELLER' | 'ADMIN';

export interface RegisterRequest {
    name: string;
    email: string;
    password: string;
    role?: Role;
}

export interface LoginRequest {
    email: string;
    password: string;
}

export interface AuthResponse {
    token: string;
    name: string;
    email: string;
    role: Role;
}